src/test/java/sfgcourse/recipeproject/converters/UnitOfMeasureCommandToUnitOfMeasureTest.java
package sfgcourse.recipeproject.converters;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import sfgcourse.recipeproject.commands.UnitOfMeasureCommand;
import sfgcourse.recipeproject.domain.UnitOfMeasure;

import static org.junit.jupiter.api.Assertions.*;

class UnitOfMeasureCommandToUnitOfMeasureTest {

    UnitOfMeasureCommandToUnitOfMeasure converter;

    @BeforeEach
    void setUp() {
        converter = new UnitOfMeasureCommandToUnitOfMeasure();
    }

    @Test
    void convertNullObject() {
        assertNull(converter.convert(null));
    }

    @Test
    void convertEmptyObject() {
        assertNotNull(converter.convert(new UnitOfMeasureCommand()));
    }

    @Test
    void convert() {
        UnitOfMeasureCommand unitOfMeasureCommand = new UnitOfMeasureCommand();
        unitOfMeasureCommand.setId(2L);
        unitOfMeasureCommand.setUom("Some uom");

        UnitOfMeasure savedUom = converter.convert(unitOfMeasureCommand);

        assertNotNull(savedUom);
        assertEquals(unitOfMeasureCommand.getId(), savedUom.getId());
        assertEquals(unitOfMeasureCommand.getUom(), savedUom.getUom());

    }
}